deployment/examples/api-que-model_as_service/pyproject.toml
[project]
name = "api-que-model-as-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.4.0",
    "fastapi>=0.115.6",
    "onnx>=1.17.0",
    "onnxruntime>=1.20.1",
    "pillow>=11.0.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]
